Implement partial current time functionality.
Incomplete, only returns startup time right now. Probably it can be done
in the main function, but probably that's the reason that the tutorial
says that `app` should have the signature `IO UTCTime ...` instead of
just `UTCTime`.

On the one hand I understand that creating the application for each
request is not quite ideal, but on the other hand it probably does that
already?

I'm not sure.